Refactor travis jobs to external scripts
For clarity, install, script and after_success scripts are now divided
in separate (short) scripts. A separate job is added outside the matrix
to perform the style checks, as those could have been bypassed during
commit time. The .gitattributes file ensures that the script files are
pushed with the correct line ending, so that they can be run on the
travis server.