Prevent the same parent twice in one pair
The parent pair should be able to perform crossover, so each parent in a
pair must be unique. I limited the population to size 3 or greater. A
population of size two only has one unique pair, which means there is
no reason to use NSGA2 selection - it is more likely something went
wrong, and I would rather raise an error.