Integrate Visualization page with Dashboard
The Visualization app was used to analyze logs after a run,
while the dashboard app would allow you to start and inspect a live run.
There was no good reason for them to be distinct.
I have commented out code responsible for aggregate plots and control.
I do not feel that this was supported cleanly and intuitively, so right
now I think it might be better not to support it at all (until I figure
out how to implement it probably and have the time to do so).
This is just the initial commit for merging them.
The plan is to unify the code with the existing Dashboard as much as
possible, to the point where a started live run can also be analyzed
with the post-run analysis.