Show data table, allow selectable target
On specifying the full path to the datafile, the arff file is loaded and
shown as a table on the HomePage. It is also possible to select a
different target column than the default (last column).
This feature does introduce visible load times even for small files,
so it's probably wise to make this optional.