Prevent excessive transfer of data to subprocesses
Data to evaluate pipelines on would before be passed on between the main
and subprocesses twice for each evaluation. The `default` class
attribute of AsyncEvaluator allows users to store default arguments
which gets passed once to the subprocess on creation. This default
argument is then always supplied when executing an async job.