Add global memory management
Replaces per-process memory management by a global memory manager.
Dynamically shrinks down the number of processes, depending on the
ratio of completed tasks to tasks terminated due to memory usage.
It (theoretically) introduces a host of other issues. Not a fan,
and hope to replace the entire Async evaluator by some mature package.