Catch TypeErrors
This surpresses the error "TypeError: an integer is required (got type
NoneType)" from within queue.get. I am not happy about this, as I do not
know the cause. I suspect the error arrives when one of the processes
that uses the queue is terminated. Either way, surpressing the error
seems to work without negative consequences. I hope to revisit this in
the future, or rather use a mature async evaluation tool/server anyway.