Move evaluation data and logs to dedicated output
This is in preparation for separated logs (progress, evaluations,
resources), which should be more user-friendly. I am not particularly
happy with the need for a `cleanup` call (could add a context manager).
Overall the logging framework gave a lot of issues, as when running unit
tests it seems like the file handle was still in use when cleanup was
called, giving errors. This lead to some of the unclean design.