Update the AsyncEvaluator, allow warm_start
Updates warm-start behavior to allow for multiple fit-calls.
Subsequent fit calls will (by default) start with the previous last
population, and run for the same time as the original fit call.

The AsyncEvaluator no longer uses a multiprocessing manager, as it
seemed to cause issues, whereas the standard multiprocessing.queue
seemed to work fine.
I have also added an extra `command` queue to allow a signal for
graceful shutdown of subprocesses instead of having to terminate the
processes. I do think this might be problematic in some place where the
pipeline evaluation is executing C-code, since then that process will
remain active until it exits the c-code.
Either way, we're currently investigating moving to `dask.distributed`
so it seems that a quick fix is probably sufficient.