Ignore 'Compute Failed' log records for timeout
These indicate a raise exception in the pipeline evaluation, but it these
can be handled when processing the future anyway. Though generally these
are expected at the end of the process where they can be ignored completely.
Printing the warnings to console might send the wrong message to the users.